move create-package to fortuneTeller routes
