--- backend/src/routes/package.routes.ts
@@ -1,9 +1,0 @@
-import express from "express"
-import { packageController } from "../controllers/package/package.controller"
-
-
-const router = express.Router()
-
-router.post("/create-package", packageController.createPackage)
-
-export default router
