+++ backend/src/services/fortuneTeller/fortuneTeller.services.ts
@@ -2,6 +2,7 @@ import { fortuneTellerRepository } from "../../repositories/fortuneTeller.reposi
 import { packageRepository } from "../../repositories/package.repository"
 import { reviewRepository } from "../../repositories/review.repository"
 import { FortuneTellerRegisterSchema, FortuneTellerAccountDetailSchema, RequestSchema } from "../../models/fortuneTeller/fortuneTeller.model"
+import { PackageSchema } from "../../models/package/package.model"
 
 export const fortuneTellerService = {
 
@@ -62,6 +63,12 @@ export const fortuneTellerService = {
       description: fortuneTeller.description,
       profilePicture: fortuneTeller.profilePicture
     }
+  },
+
+  createPackage: async (packageFortuneTeller: PackageSchema) => {
+    const isSuccess = await packageRepository.createPackage(packageFortuneTeller)
+    return { success: isSuccess, message: (isSuccess) ? "success" : "error to create package" } 
+    
   },
 
   getPackageByFortuneTellerId: async (fortuneTellerId: string) => {
